pyproject.toml
[project]
name = "flow-JKO"
version = "0.1"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "torchdiffeq>=0.2.5",
    "wandb>=0.21.3",
]
